Delay vectorization of paths until necessary
This fixes a bug that transforms were only parsed after the path
data had already been evaluated with the wrong transformation.

This also makes the naming between polylines/paths/subpaths/etc.
somewhat more consistent.

Also fixes a bug in the tiling algorithm (casting to int was used
as a floor operation, which is obviously incorrect when the
numbers can be negative.

Fixes #5, #3